# Avoid branching on ntz shift