Since Go sets values shifted outside their range, it needs to add a branch on variable shifts.

Since we know this value cannot be outside of 0->63 we provide the compiler a hint that this cannot happen.

```
benchmark                                                            old ns/op     new ns/op     delta
BenchmarkCompressDeltaBinPackInt64/nBits=0,_ntz=0,_sign=0-32         438           434           -1.12%
BenchmarkCompressDeltaBinPackInt64/nBits=0,_ntz=0,_sign=1-32         441           432           -1.91%
BenchmarkCompressDeltaBinPackInt64/nBits=0,_ntz=31,_sign=0-32        442           432           -2.22%
BenchmarkCompressDeltaBinPackInt64/nBits=0,_ntz=31,_sign=1-32        440           432           -1.89%
BenchmarkCompressDeltaBinPackInt64/nBits=0,_ntz=62,_sign=0-32        442           433           -1.95%
BenchmarkCompressDeltaBinPackInt64/nBits=0,_ntz=62,_sign=1-32        431           434           +0.65%
BenchmarkCompressDeltaBinPackInt64/nBits=1,_ntz=0,_sign=0-32         578           556           -3.86%
BenchmarkCompressDeltaBinPackInt64/nBits=1,_ntz=0,_sign=1-32         536           508           -5.26%
BenchmarkCompressDeltaBinPackInt64/nBits=1,_ntz=31,_sign=0-32        530           501           -5.43%
BenchmarkCompressDeltaBinPackInt64/nBits=1,_ntz=31,_sign=1-32        531           506           -4.62%
BenchmarkCompressDeltaBinPackInt64/nBits=1,_ntz=62,_sign=0-32        594           554           -6.80%
BenchmarkCompressDeltaBinPackInt64/nBits=1,_ntz=62,_sign=1-32        595           550           -7.50%
BenchmarkCompressDeltaBinPackInt64/nBits=2,_ntz=0,_sign=0-32         575           561           -2.47%
BenchmarkCompressDeltaBinPackInt64/nBits=2,_ntz=0,_sign=1-32         530           504           -4.85%
BenchmarkCompressDeltaBinPackInt64/nBits=2,_ntz=31,_sign=0-32        532           507           -4.74%
BenchmarkCompressDeltaBinPackInt64/nBits=2,_ntz=31,_sign=1-32        530           505           -4.77%
BenchmarkCompressDeltaBinPackInt64/nBits=2,_ntz=62,_sign=0-32        898           749           -16.59%
BenchmarkCompressDeltaBinPackInt64/nBits=2,_ntz=62,_sign=1-32        458           450           -1.85%
BenchmarkCompressDeltaBinPackInt64/nBits=3,_ntz=0,_sign=0-32         584           565           -3.37%
BenchmarkCompressDeltaBinPackInt64/nBits=3,_ntz=0,_sign=1-32         531           505           -4.86%
BenchmarkCompressDeltaBinPackInt64/nBits=3,_ntz=31,_sign=0-32        535           505           -5.48%
BenchmarkCompressDeltaBinPackInt64/nBits=3,_ntz=31,_sign=1-32        532           509           -4.27%
BenchmarkCompressDeltaBinPackInt64/nBits=4,_ntz=0,_sign=0-32         581           554           -4.77%
BenchmarkCompressDeltaBinPackInt64/nBits=4,_ntz=0,_sign=1-32         525           502           -4.27%
BenchmarkCompressDeltaBinPackInt64/nBits=4,_ntz=31,_sign=0-32        528           500           -5.29%
BenchmarkCompressDeltaBinPackInt64/nBits=4,_ntz=31,_sign=1-32        525           502           -4.42%
BenchmarkCompressDeltaBinPackInt64/nBits=5,_ntz=0,_sign=0-32         614           572           -6.80%
BenchmarkCompressDeltaBinPackInt64/nBits=5,_ntz=0,_sign=1-32         547           512           -6.31%
BenchmarkCompressDeltaBinPackInt64/nBits=5,_ntz=31,_sign=0-32        540           512           -5.15%
BenchmarkCompressDeltaBinPackInt64/nBits=5,_ntz=31,_sign=1-32        537           512           -4.71%
BenchmarkCompressDeltaBinPackInt64/nBits=6,_ntz=0,_sign=0-32         588           580           -1.33%
BenchmarkCompressDeltaBinPackInt64/nBits=6,_ntz=0,_sign=1-32         533           515           -3.30%
BenchmarkCompressDeltaBinPackInt64/nBits=6,_ntz=31,_sign=0-32        534           514           -3.78%
BenchmarkCompressDeltaBinPackInt64/nBits=6,_ntz=31,_sign=1-32        534           518           -3.13%
BenchmarkCompressDeltaBinPackInt64/nBits=7,_ntz=0,_sign=0-32         603           579           -3.91%
BenchmarkCompressDeltaBinPackInt64/nBits=7,_ntz=0,_sign=1-32         551           513           -6.88%
BenchmarkCompressDeltaBinPackInt64/nBits=7,_ntz=31,_sign=0-32        546           518           -5.10%
BenchmarkCompressDeltaBinPackInt64/nBits=7,_ntz=31,_sign=1-32        552           513           -7.19%
BenchmarkCompressDeltaBinPackInt64/nBits=8,_ntz=0,_sign=0-32         579           552           -4.68%
BenchmarkCompressDeltaBinPackInt64/nBits=8,_ntz=0,_sign=1-32         533           504           -5.57%
BenchmarkCompressDeltaBinPackInt64/nBits=8,_ntz=31,_sign=0-32        529           501           -5.16%
BenchmarkCompressDeltaBinPackInt64/nBits=8,_ntz=31,_sign=1-32        524           499           -4.70%
BenchmarkCompressDeltaBinPackInt64/nBits=9,_ntz=0,_sign=0-32         613           579           -5.55%
BenchmarkCompressDeltaBinPackInt64/nBits=9,_ntz=0,_sign=1-32         558           512           -8.10%
BenchmarkCompressDeltaBinPackInt64/nBits=9,_ntz=31,_sign=0-32        546           518           -5.20%
BenchmarkCompressDeltaBinPackInt64/nBits=9,_ntz=31,_sign=1-32        542           522           -3.80%
BenchmarkCompressDeltaBinPackInt64/nBits=10,_ntz=0,_sign=0-32        601           570           -5.14%
BenchmarkCompressDeltaBinPackInt64/nBits=10,_ntz=0,_sign=1-32        547           518           -5.32%
BenchmarkCompressDeltaBinPackInt64/nBits=10,_ntz=31,_sign=0-32       539           516           -4.31%
BenchmarkCompressDeltaBinPackInt64/nBits=10,_ntz=31,_sign=1-32       547           514           -5.94%
BenchmarkCompressDeltaBinPackInt64/nBits=11,_ntz=0,_sign=0-32        608           582           -4.24%
BenchmarkCompressDeltaBinPackInt64/nBits=11,_ntz=0,_sign=1-32        560           535           -4.57%
BenchmarkCompressDeltaBinPackInt64/nBits=11,_ntz=31,_sign=0-32       568           532           -6.44%
BenchmarkCompressDeltaBinPackInt64/nBits=11,_ntz=31,_sign=1-32       568           528           -7.02%
BenchmarkCompressDeltaBinPackInt64/nBits=12,_ntz=0,_sign=0-32        593           576           -2.82%
BenchmarkCompressDeltaBinPackInt64/nBits=12,_ntz=0,_sign=1-32        542           517           -4.54%
BenchmarkCompressDeltaBinPackInt64/nBits=12,_ntz=31,_sign=0-32       542           512           -5.43%
BenchmarkCompressDeltaBinPackInt64/nBits=12,_ntz=31,_sign=1-32       538           513           -4.68%
BenchmarkCompressDeltaBinPackInt64/nBits=13,_ntz=0,_sign=0-32        650           583           -10.26%
BenchmarkCompressDeltaBinPackInt64/nBits=13,_ntz=0,_sign=1-32        554           521           -5.87%
BenchmarkCompressDeltaBinPackInt64/nBits=13,_ntz=31,_sign=0-32       557           520           -6.68%
BenchmarkCompressDeltaBinPackInt64/nBits=13,_ntz=31,_sign=1-32       553           521           -5.86%
BenchmarkCompressDeltaBinPackInt64/nBits=14,_ntz=0,_sign=0-32        616           604           -2.03%
BenchmarkCompressDeltaBinPackInt64/nBits=14,_ntz=0,_sign=1-32        558           527           -5.42%
BenchmarkCompressDeltaBinPackInt64/nBits=14,_ntz=31,_sign=0-32       550           525           -4.53%
BenchmarkCompressDeltaBinPackInt64/nBits=14,_ntz=31,_sign=1-32       548           527           -3.81%
BenchmarkCompressDeltaBinPackInt64/nBits=15,_ntz=0,_sign=0-32        621           595           -4.27%
BenchmarkCompressDeltaBinPackInt64/nBits=15,_ntz=0,_sign=1-32        557           526           -5.55%
BenchmarkCompressDeltaBinPackInt64/nBits=15,_ntz=31,_sign=0-32       564           529           -6.19%
BenchmarkCompressDeltaBinPackInt64/nBits=15,_ntz=31,_sign=1-32       555           525           -5.42%
BenchmarkCompressDeltaBinPackInt64/nBits=16,_ntz=0,_sign=0-32        599           554           -7.41%
BenchmarkCompressDeltaBinPackInt64/nBits=16,_ntz=0,_sign=1-32        521           501           -3.95%
BenchmarkCompressDeltaBinPackInt64/nBits=16,_ntz=31,_sign=0-32       520           501           -3.52%
BenchmarkCompressDeltaBinPackInt64/nBits=16,_ntz=31,_sign=1-32       532           502           -5.53%
BenchmarkCompressDeltaBinPackInt64/nBits=17,_ntz=0,_sign=0-32        662           615           -7.03%
BenchmarkCompressDeltaBinPackInt64/nBits=17,_ntz=0,_sign=1-32        563           530           -5.92%
BenchmarkCompressDeltaBinPackInt64/nBits=17,_ntz=31,_sign=0-32       564           526           -6.64%
BenchmarkCompressDeltaBinPackInt64/nBits=17,_ntz=31,_sign=1-32       561           525           -6.35%
BenchmarkCompressDeltaBinPackInt64/nBits=18,_ntz=0,_sign=0-32        624           626           +0.34%
BenchmarkCompressDeltaBinPackInt64/nBits=18,_ntz=0,_sign=1-32        566           530           -6.50%
BenchmarkCompressDeltaBinPackInt64/nBits=18,_ntz=31,_sign=0-32       560           532           -4.95%
BenchmarkCompressDeltaBinPackInt64/nBits=18,_ntz=31,_sign=1-32       568           527           -7.15%
BenchmarkCompressDeltaBinPackInt64/nBits=19,_ntz=0,_sign=0-32        632           606           -4.08%
BenchmarkCompressDeltaBinPackInt64/nBits=19,_ntz=0,_sign=1-32        569           532           -6.49%
BenchmarkCompressDeltaBinPackInt64/nBits=19,_ntz=31,_sign=0-32       567           531           -6.33%
BenchmarkCompressDeltaBinPackInt64/nBits=19,_ntz=31,_sign=1-32       570           530           -7.10%
BenchmarkCompressDeltaBinPackInt64/nBits=20,_ntz=0,_sign=0-32        648           593           -8.58%
BenchmarkCompressDeltaBinPackInt64/nBits=20,_ntz=0,_sign=1-32        557           525           -5.75%
BenchmarkCompressDeltaBinPackInt64/nBits=20,_ntz=31,_sign=0-32       564           529           -6.22%
BenchmarkCompressDeltaBinPackInt64/nBits=20,_ntz=31,_sign=1-32       566           529           -6.57%
BenchmarkCompressDeltaBinPackInt64/nBits=21,_ntz=0,_sign=0-32        637           621           -2.54%
BenchmarkCompressDeltaBinPackInt64/nBits=21,_ntz=0,_sign=1-32        574           536           -6.52%
BenchmarkCompressDeltaBinPackInt64/nBits=21,_ntz=31,_sign=0-32       571           534           -6.51%
BenchmarkCompressDeltaBinPackInt64/nBits=21,_ntz=31,_sign=1-32       574           534           -6.97%
BenchmarkCompressDeltaBinPackInt64/nBits=22,_ntz=0,_sign=0-32        639           607           -5.04%
BenchmarkCompressDeltaBinPackInt64/nBits=22,_ntz=0,_sign=1-32        574           538           -6.32%
BenchmarkCompressDeltaBinPackInt64/nBits=22,_ntz=31,_sign=0-32       568           542           -4.53%
BenchmarkCompressDeltaBinPackInt64/nBits=22,_ntz=31,_sign=1-32       566           540           -4.61%
BenchmarkCompressDeltaBinPackInt64/nBits=23,_ntz=0,_sign=0-32        647           622           -3.88%
BenchmarkCompressDeltaBinPackInt64/nBits=23,_ntz=0,_sign=1-32        578           540           -6.64%
BenchmarkCompressDeltaBinPackInt64/nBits=23,_ntz=31,_sign=0-32       574           538           -6.26%
BenchmarkCompressDeltaBinPackInt64/nBits=23,_ntz=31,_sign=1-32       573           533           -7.06%
BenchmarkCompressDeltaBinPackInt64/nBits=24,_ntz=0,_sign=0-32        620           582           -6.13%
BenchmarkCompressDeltaBinPackInt64/nBits=24,_ntz=0,_sign=1-32        549           522           -4.79%
BenchmarkCompressDeltaBinPackInt64/nBits=24,_ntz=31,_sign=0-32       548           527           -3.94%
BenchmarkCompressDeltaBinPackInt64/nBits=24,_ntz=31,_sign=1-32       549           522           -4.90%
BenchmarkCompressDeltaBinPackInt64/nBits=25,_ntz=0,_sign=0-32        654           618           -5.46%
BenchmarkCompressDeltaBinPackInt64/nBits=25,_ntz=0,_sign=1-32        582           541           -7.08%
BenchmarkCompressDeltaBinPackInt64/nBits=25,_ntz=31,_sign=0-32       583           538           -7.64%
BenchmarkCompressDeltaBinPackInt64/nBits=25,_ntz=31,_sign=1-32       579           539           -6.94%
BenchmarkCompressDeltaBinPackInt64/nBits=26,_ntz=0,_sign=0-32        645           614           -4.88%
BenchmarkCompressDeltaBinPackInt64/nBits=26,_ntz=0,_sign=1-32        585           545           -6.71%
BenchmarkCompressDeltaBinPackInt64/nBits=26,_ntz=31,_sign=0-32       580           545           -6.13%
BenchmarkCompressDeltaBinPackInt64/nBits=26,_ntz=31,_sign=1-32       585           546           -6.58%
BenchmarkCompressDeltaBinPackInt64/nBits=27,_ntz=0,_sign=0-32        661           623           -5.69%
BenchmarkCompressDeltaBinPackInt64/nBits=27,_ntz=0,_sign=1-32        590           542           -8.10%
BenchmarkCompressDeltaBinPackInt64/nBits=27,_ntz=31,_sign=0-32       580           545           -6.05%
BenchmarkCompressDeltaBinPackInt64/nBits=27,_ntz=31,_sign=1-32       592           539           -9.00%
BenchmarkCompressDeltaBinPackInt64/nBits=28,_ntz=0,_sign=0-32        641           617           -3.65%
BenchmarkCompressDeltaBinPackInt64/nBits=28,_ntz=0,_sign=1-32        576           544           -5.47%
BenchmarkCompressDeltaBinPackInt64/nBits=28,_ntz=31,_sign=0-32       575           550           -4.38%
BenchmarkCompressDeltaBinPackInt64/nBits=28,_ntz=31,_sign=1-32       573           544           -5.15%
BenchmarkCompressDeltaBinPackInt64/nBits=29,_ntz=0,_sign=0-32        690           620           -10.15%
BenchmarkCompressDeltaBinPackInt64/nBits=29,_ntz=0,_sign=1-32        597           542           -9.18%
BenchmarkCompressDeltaBinPackInt64/nBits=29,_ntz=31,_sign=0-32       594           547           -7.86%
BenchmarkCompressDeltaBinPackInt64/nBits=29,_ntz=31,_sign=1-32       591           545           -7.75%
BenchmarkCompressDeltaBinPackInt64/nBits=30,_ntz=0,_sign=0-32        688           624           -9.24%
BenchmarkCompressDeltaBinPackInt64/nBits=30,_ntz=0,_sign=1-32        592           547           -7.60%
BenchmarkCompressDeltaBinPackInt64/nBits=30,_ntz=31,_sign=0-32       591           551           -6.67%
BenchmarkCompressDeltaBinPackInt64/nBits=30,_ntz=31,_sign=1-32       590           551           -6.68%
BenchmarkCompressDeltaBinPackInt64/nBits=31,_ntz=0,_sign=0-32        689           622           -9.71%
BenchmarkCompressDeltaBinPackInt64/nBits=31,_ntz=0,_sign=1-32        593           550           -7.20%
BenchmarkCompressDeltaBinPackInt64/nBits=31,_ntz=31,_sign=0-32       593           551           -7.06%
BenchmarkCompressDeltaBinPackInt64/nBits=31,_ntz=31,_sign=1-32       590           550           -6.78%
BenchmarkCompressDeltaBinPackInt64/nBits=32,_ntz=0,_sign=0-32        567           549           -3.05%
BenchmarkCompressDeltaBinPackInt64/nBits=32,_ntz=0,_sign=1-32        511           498           -2.49%
BenchmarkCompressDeltaBinPackInt64/nBits=32,_ntz=31,_sign=0-32       513           502           -2.18%
BenchmarkCompressDeltaBinPackInt64/nBits=32,_ntz=31,_sign=1-32       511           498           -2.54%
BenchmarkCompressDeltaBinPackInt64/nBits=33,_ntz=0,_sign=0-32        735           637           -13.33%
BenchmarkCompressDeltaBinPackInt64/nBits=33,_ntz=0,_sign=1-32        593           557           -6.09%
BenchmarkCompressDeltaBinPackInt64/nBits=33,_ntz=31,_sign=0-32       893           744           -16.74%
BenchmarkCompressDeltaBinPackInt64/nBits=33,_ntz=31,_sign=1-32       450           451           +0.22%
BenchmarkCompressDeltaBinPackInt64/nBits=34,_ntz=0,_sign=0-32        700           627           -10.50%
BenchmarkCompressDeltaBinPackInt64/nBits=34,_ntz=0,_sign=1-32        590           552           -6.46%
BenchmarkCompressDeltaBinPackInt64/nBits=35,_ntz=0,_sign=0-32        725           630           -13.15%
BenchmarkCompressDeltaBinPackInt64/nBits=35,_ntz=0,_sign=1-32        612           554           -9.43%
BenchmarkCompressDeltaBinPackInt64/nBits=36,_ntz=0,_sign=0-32        702           635           -9.53%
BenchmarkCompressDeltaBinPackInt64/nBits=36,_ntz=0,_sign=1-32        585           552           -5.63%
BenchmarkCompressDeltaBinPackInt64/nBits=37,_ntz=0,_sign=0-32        702           632           -9.93%
BenchmarkCompressDeltaBinPackInt64/nBits=37,_ntz=0,_sign=1-32        600           559           -6.95%
BenchmarkCompressDeltaBinPackInt64/nBits=38,_ntz=0,_sign=0-32        728           636           -12.71%
BenchmarkCompressDeltaBinPackInt64/nBits=38,_ntz=0,_sign=1-32        604           560           -7.33%
BenchmarkCompressDeltaBinPackInt64/nBits=39,_ntz=0,_sign=0-32        730           644           -11.91%
BenchmarkCompressDeltaBinPackInt64/nBits=39,_ntz=0,_sign=1-32        607           562           -7.46%
BenchmarkCompressDeltaBinPackInt64/nBits=40,_ntz=0,_sign=0-32        645           626           -2.98%
BenchmarkCompressDeltaBinPackInt64/nBits=40,_ntz=0,_sign=1-32        574           545           -5.02%
BenchmarkCompressDeltaBinPackInt64/nBits=41,_ntz=0,_sign=0-32        730           698           -4.33%
BenchmarkCompressDeltaBinPackInt64/nBits=41,_ntz=0,_sign=1-32        616           563           -8.51%
BenchmarkCompressDeltaBinPackInt64/nBits=42,_ntz=0,_sign=0-32        781           669           -14.37%
BenchmarkCompressDeltaBinPackInt64/nBits=42,_ntz=0,_sign=1-32        621           572           -7.83%
BenchmarkCompressDeltaBinPackInt64/nBits=43,_ntz=0,_sign=0-32        754           649           -13.97%
BenchmarkCompressDeltaBinPackInt64/nBits=43,_ntz=0,_sign=1-32        624           569           -8.95%
BenchmarkCompressDeltaBinPackInt64/nBits=44,_ntz=0,_sign=0-32        738           644           -12.79%
BenchmarkCompressDeltaBinPackInt64/nBits=44,_ntz=0,_sign=1-32        617           559           -9.45%
BenchmarkCompressDeltaBinPackInt64/nBits=45,_ntz=0,_sign=0-32        815           691           -15.16%
BenchmarkCompressDeltaBinPackInt64/nBits=45,_ntz=0,_sign=1-32        652           572           -12.37%
BenchmarkCompressDeltaBinPackInt64/nBits=46,_ntz=0,_sign=0-32        789           668           -15.27%
BenchmarkCompressDeltaBinPackInt64/nBits=46,_ntz=0,_sign=1-32        662           574           -13.36%
BenchmarkCompressDeltaBinPackInt64/nBits=47,_ntz=0,_sign=0-32        778           671           -13.82%
BenchmarkCompressDeltaBinPackInt64/nBits=47,_ntz=0,_sign=1-32        636           572           -10.07%
BenchmarkCompressDeltaBinPackInt64/nBits=48,_ntz=0,_sign=0-32        653           630           -3.54%
BenchmarkCompressDeltaBinPackInt64/nBits=48,_ntz=0,_sign=1-32        577           543           -5.84%
BenchmarkCompressDeltaBinPackInt64/nBits=49,_ntz=0,_sign=0-32        813           712           -12.47%
BenchmarkCompressDeltaBinPackInt64/nBits=49,_ntz=0,_sign=1-32        641           575           -10.33%
BenchmarkCompressDeltaBinPackInt64/nBits=50,_ntz=0,_sign=0-32        1013          718           -29.08%
BenchmarkCompressDeltaBinPackInt64/nBits=50,_ntz=0,_sign=1-32        683           577           -15.58%
BenchmarkCompressDeltaBinPackInt64/nBits=51,_ntz=0,_sign=0-32        860           740           -13.95%
BenchmarkCompressDeltaBinPackInt64/nBits=51,_ntz=0,_sign=1-32        642           580           -9.68%
BenchmarkCompressDeltaBinPackInt64/nBits=52,_ntz=0,_sign=0-32        736           702           -4.73%
BenchmarkCompressDeltaBinPackInt64/nBits=52,_ntz=0,_sign=1-32        617           580           -5.85%
BenchmarkCompressDeltaBinPackInt64/nBits=53,_ntz=0,_sign=0-32        841           693           -17.63%
BenchmarkCompressDeltaBinPackInt64/nBits=53,_ntz=0,_sign=1-32        661           586           -11.36%
BenchmarkCompressDeltaBinPackInt64/nBits=54,_ntz=0,_sign=0-32        789           711           -9.86%
BenchmarkCompressDeltaBinPackInt64/nBits=54,_ntz=0,_sign=1-32        667           589           -11.62%
BenchmarkCompressDeltaBinPackInt64/nBits=55,_ntz=0,_sign=0-32        877           720           -17.84%
BenchmarkCompressDeltaBinPackInt64/nBits=55,_ntz=0,_sign=1-32        666           584           -12.26%
BenchmarkCompressDeltaBinPackInt64/nBits=56,_ntz=0,_sign=0-32        756           692           -8.52%
…