@klauspost 2023-03-27
Since you are on an arm64 platform maybe check if this is a regression for you.

If so, we can use platform defined constants, like I do in my deflate: [amd64](https://github.com/klauspost/compress/blob/master/flate/regmask_amd64.go), [others](https://github.com/klauspost/compress/blob/master/flate/regmask_other.go). 

I have checked that "others" creates nops, though I don't currently have it defined for ints, only uints.